Fix initialize and update tests
