+++ sense2vec/component.py
@@ -1,6 +1,7 @@
-from typing import Tuple, Union, List, Dict
+from typing import Tuple, Union, List, Dict, Callable, Iterable, Optional
 from spacy.language import Language
 from spacy.tokens import Doc, Token, Span
+from spacy.training import Example
 from spacy.vocab import Vocab
 from spacy.util import SimpleFrozenDict
 from pathlib import Path
@@ -214,6 +215,24 @@ class Sense2VecComponent(object):
         """
         key = self.s2v_key(obj)
         return obj.doc._._s2v.get_other_senses(key)
+
+    def initialize(
+        self,
+        get_examples: Callable[[], Iterable[Example]],
+        *,
+        nlp: Optional[Language] = None,
+        data_path: Optional[str] = None
+    ):
+        """Initialize the component and load in data. Can be used to add the
+        component with vectors to a pipeline before training.
+
+        get_examples (Callable[[], Iterable[Example]]): Function that
+            returns a representative sample of gold-standard Example objects.
+        nlp (Language): The current nlp object the component is part of.
+        data_path (Optional[str]): Optional path to sense2vec model.
+        """
+        if data_path is not None:
+            self.from_disk(data_path)
 
     def to_bytes(self) -> bytes:
         """Serialize the component to a bytestring.
