+++ setup.cfg
@@ -36,7 +36,7 @@ install_requires =
 
 [options.entry_points]
 spacy_factories =
-    sense2vec = sense2vec:make_sense2vec
+    sense2vec = sense2vec:component.make_sense2vec
 prodigy_recipes =
     sense2vec.teach = sense2vec:prodigy_recipes.teach
     sens2vec.to-patterns = sense2vec:prodigy_recipes.to_patterns
